requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
kafka-python==2.2.15
python-dotenv==1.1.1
requests==2.32.5
urllib3==2.5.0
